Refactor match player into presentational and route+fetch components
