+++ src/models/Telemetry.worker.js
@@ -1,10 +1,14 @@
 import parseTelemetry from './Telemetry.parser.js'
 
 async function handleMessage({ data: { match, focusedPlayer } }) {
-    const res = await fetch(match.telemetryUrl)
-    const telemetryData = await res.json()
-    const { state, globalState } = parseTelemetry(match, telemetryData, focusedPlayer)
-    postMessage({ state, globalState })
+    try {
+        const res = await fetch(match.telemetryUrl)
+        const telemetryData = await res.json()
+        const { state, globalState } = parseTelemetry(match, telemetryData, focusedPlayer)
+        postMessage({ success: true, state, globalState })
+    } catch (error) {
+        postMessage({ success: false, error: error.message })
+    }
 }
 
 self.addEventListener('message', handleMessage)
